+++ hvad/tests/admin.py
@@ -1,5 +1,4 @@
 # -*- coding: utf-8 -*-
-import django
 from django.contrib import admin
 from django.contrib.contenttypes.models import ContentType
 from django.core.urlresolvers import reverse